refactor: create `utils/config` folder - Move `dbConfig` to `utils\config\db.ts`. - Rename `utils\db\config.ts` to be `utils\db\provider.ts`
