+++ prisma/migrate.ts
@@ -1,5 +1,6 @@
+import { dbConfig } from '~/utils/config/db';
 import { validateDBConfig } from '~/utils/db';
-import { dbConfig, dbProvider } from '~/utils/db/config';
+import { dbProvider } from '~/utils/db/provider';
 import { runOrFail } from '../utils/helpers';
 
 const main = async () => {
