+++ utils/config/db.ts
@@ -1,0 +1,6 @@
+import 'dotenv/config';
+
+export const dbConfig = {
+  provider: process.env.DB_PROVIDER,
+  url: process.env.DB_URL,
+} as const;
