+++ utils/db/provider.ts
@@ -1,10 +1,3 @@
-import 'dotenv/config';
-
-export const dbConfig = {
-  provider: process.env.DB_PROVIDER,
-  url: process.env.DB_URL,
-} as const;
-
 export const dbProvider = {
   mongodb: 'mongodb',
   postgresql: 'postgresql',
